% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.r
\name{dev_asp}
\alias{dev_asp}
\title{Get graphics device aspect ratio}
\usage{
dev_asp(type = c("w2h", "h2w"))
}
\arguments{
\item{type}{One of width-to-height (w2h) or height-to-width (h2w).
Defaults to "w2h".}
}
\description{
Get graphics device aspect ratio
}
