% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{dt}
\alias{dt}
\alias{d}
\title{Alias for data.table's \code{[}}
\usage{
dt(
  x,
  i,
  j,
  by,
  keyby,
  with = TRUE,
  nomatch = getOption("datatable.nomatch", NA),
  mult = "all",
  roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE, TRUE) else
    c(TRUE, FALSE),
  which = FALSE,
  .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"),
  drop = NULL,
  on = NULL
)

d(
  x,
  i,
  j,
  by,
  keyby,
  with = TRUE,
  nomatch = getOption("datatable.nomatch", NA),
  mult = "all",
  roll = FALSE,
  rollends = if (roll == "nearest") c(TRUE, TRUE) else if (roll >= 0) c(FALSE, TRUE) else
    c(TRUE, FALSE),
  which = FALSE,
  .SDcols,
  verbose = getOption("datatable.verbose"),
  allow.cartesian = getOption("datatable.allow.cartesian"),
  drop = NULL,
  on = NULL
)
}
\description{
Aliases for the \code{[} function in \code{data.table}, for manipulating
data tables using R's native pipe operator.
}
\details{
Not my idea (see: eliocamp's comments):
https://github.com/Rdatatable/data.table/issues/4872
}
\section{Functions}{
\itemize{
\item \code{d()}: d

}}
