% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{expand_grid}
\alias{expand_grid}
\title{Convert output of expand.grid to data.table}
\usage{
expand_grid(..., sorted = TRUE, unique = FALSE)
}
\arguments{
\item{...}{Pass arguments to \code{expand.grid()}}
}
\description{
An alias for data.table's \code{CJ()}, which is like \code{expand.grid()} but returns a \code{data.table}
}
