% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numbers.r
\name{fbig}
\alias{fbig}
\title{Commas in big numbers}
\usage{
fbig(...)
}
\arguments{
\item{...}{Feed arguments to \code{format()}.}
}
\description{
A wrapper around \code{format()} that sets \code{big.mark =
  ","}.  Convenient for repeated use in manuscripts, tables, and other
outputs meant for human consumption.
}
