% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{getnames}
\alias{getnames}
\title{Retrieve column names}
\usage{
getnames(pat, data)
}
\arguments{
\item{pat}{A regular expression defining the pattern(s) used to retrieve column names.}

\item{data}{A data frame from which to retrieve the column names.}
}
\description{
Helper function that returns column names from a data frame based on a regular expression.
}
