% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install-packages.r
\name{install_packages}
\alias{install_packages}
\title{Install commonly used R packages}
\usage{
install_packages(main = TRUE, spatial = FALSE, quickdag = FALSE)
}
\arguments{
\item{main}{A Boolean indicating whether to install the main list of packages. Default is \code{TRUE}.}

\item{spatial}{A Boolean indicating whether to install spatial packages. Default is \code{FALSE}.}

\item{quickdag}{A Boolean indicating whether to install the \code{quickdag} package. Default is \code{FALSE}.}
}
\value{
No objects are returned. The function simply installs a list of R packages.
}
\description{
This function installs a list of R packages I use frequently. Primarily for use with a fresh or upgraded R installation. Note that the function name is not exported into the global environment, in order to avoid unintentional use that may affect ongoing projects.
}
\examples{
install.packages("remotes")
remotes::install_github("jrgant/gantenmisc")
gantenmisc::install_packages()
}
