% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatable.r
\name{expand_grid}
\alias{expand_grid}
\title{Convert output of expand.grid to data.table}
\usage{
expand_grid(...)
}
\arguments{
\item{...}{Pass arguments to \code{expand.grid()}}
}
\description{
A wrapper around \code{expand.grid()} that outputs a \code{data.table} instead of a vanilla \code{data.frame}.
}
